refactor: [#68] Split code into separate file
