+++ MoonCrystal/MoonCrystal/UserProfile/UserProfileDetailView.swift
@@ -14,7 +14,6 @@ struct UserProfileDetailView: View {
     @Environment(\.presentationMode) var presentationMode
     
     @State var keyboardHeight : CGFloat = 0
-    
     @State private var isEditing = false
     @State private var userProfileInputData = UserProfileInputModel()
     
@@ -46,7 +45,6 @@ struct UserProfileDetailView: View {
         .padding(.bottom, 0)
         .navigationTitle(isEditing ? "프로필 수정" : "프로필")
         .navigationBarBackButtonHidden(true)
-        
         .toolbar {
             ToolbarItem(placement: .topBarLeading) {
                 backButton
@@ -146,38 +144,5 @@ struct UserProfileDetailView: View {
                 .font(.system(size: 16, weight: .regular))
                 .foregroundColor(isEditing ? .pink300 : .gray700)
         }
-    }
-    
-}
-
-struct KeyboardProvider : ViewModifier {
-    
-    //키보드 높이값
-    var keyboardHeight: Binding<CGFloat>
-    
-    func body(content: Content) -> some View {
-        content
-        //키보드 올라가기 직전 노티를 받으면 나오는 객체
-            .onReceive(NotificationCenter.default.publisher(for: UIResponder.keyboardWillShowNotification),
-                       perform: { notification in
-                guard let userInfo = notification.userInfo,
-                      let keyboardRect = userInfo[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect else { return }
-                
-                //키보드 높이값 . 바인딩 원본 객체 연결 -> 전달
-                self.keyboardHeight.wrappedValue = keyboardRect.height
-                
-            })
-        //키보드 닫기 전 보내는 노티 받으면 실행
-            .onReceive(NotificationCenter.default.publisher(for: UIResponder.keyboardWillHideNotification),
-                       perform: { _ in
-                //키보드 높이값 0으로 변경
-                self.keyboardHeight.wrappedValue = 0
-            })
-    }
-}
-
-public extension View {
-    func keyboardHeight(_ state: Binding<CGFloat>) -> some View {
-        self.modifier(KeyboardProvider(keyboardHeight: state))
     }
 }
